1.3 stage 0 generator is mostly working
I don't have any kind of test harness set up yet, and I
don't have an actual binary at this time. But the actual
generation of assembly code for stage 1 (as a string) is working,
and I'm able to assemble it into a running binary.

I've added a docker-compose setup so that it's easy to get
a linux environment where things run, and I've added some commands
to the readme for using gcc.