Finish stage 5
There are unit tests of parsing, and unit tests of the
new Context model. In addition, all integ tests of stage 5
are passing.

This was a very big commit, because I wanted the context
handling to be able to generalize when we get to inner
scope blocks and functions.

Unfortunately, emacs broke during this stage - it now
reports a lot of errors in both the ast and context
modules.

All of these errors result from errors of the form:
  "Constraints are not satisfied in this type."
    "Type xxx should be an instance of xxx"

None of the errors are reported by dune, which means that
somehow the emacs merlin integration and dune are getting out
of sync :(

The problem doesn't appear to be my code per se, so I'll probably
put this off for now. I did try running make clean to be sure
emacs didn't cache an old merlin file or something - did not help.