Stage 8: loops and loop control statements
There is a test suite for parsing:
 - for loops with no declaration
 - for loops with a declaration
 - while loops
 - do-while loops
 - break and continue statements

In addition, the lexing and parsing was refactored slightly so
that annotations take a string input, builtin types have a different
token form, and we can correctly parse definitions that involve
user-defined types (which have an ID token). There are unit test
of this for both
 - variable definitiions
 - function definitions

We also refactored the context a bit to handle loops, which requires
maintaining a stack of current loops so that we can properly handle
break and continue statements. As part of this, we refactored
the "scope level" into "block scope" so that we could have
 - `enter_block` / `exit_block` functions, and then by analogy we can have
 - `enter_loop` / `exit_loop` functions.
The old code all works, and the new context code is unit tested.

There are no integ tests of code generation, but all integ tests pass.
The technique used is to create a single, generic loop-assembly-code
function, and convert all 4 types of loops (while, do, for/decl,
for/expr) into it.

All integ tests thru stage 8 pass.