Start stage 9 (fcalls, fdefs). Fix bug
This commit is a partway point on stage 9, which I'm going
to save as a non-WIP commit because of an important
improvement to Context.ml.

The first part of this commit is that we have parsing
working for function definitions with arguments and function
calls. This is unit tested. Also, fixed a bug where functions
were required to have block statements

The second part of the commit is that I went through Context.ml
carefully:
 - I rewrote all of the big block comments explaining stack
   management (some of them were not quite up to date with my
   current understanding)
 - I fixed an important conceptual bug (the code would have worked,
   but wasted stack space) where I improperly propagated ebp offsets
   from the arg-level block scope to lower block scopes.
 - I fixed some nasty issues where things that should be optional
   were not, and added some utility methods to check the "state"
   of a context (top-level, in a function definition, in a block scope
   where we can add locals) so that we'll get usable errors if we
   enter a bad state
 - Separated the concepts of function declaration and function
   definition; declaration adds the function to the function map.
   Definition involves calls to `enter_fn` / `exit_fn`.
 - I added a `find_fn_info` function, with unit test