Finish step 9: basic function defs / calls
The compiler now passes all integration tests.

The code in some of our validations got pretty messy during this
step; at this point some of the Context code badly needs a refactor,
which is may never get since this commit completes the nqcc project.

Also, in the process I realized that some of the code was only
compiling because of opam-version-depenent stuff that made it
fail in Linux; I fixed all these errors, although at the moment
I'm still having issues installing ounit in docker /shrug